Move TargetReference to it's own file
